Avoid creation of .irbrc if directory isn't writeable
Fixes #212

+++ resources/files/irbrc_predefiner.rb
@@ -2,7 +2,7 @@ require 'irb'
 
 irbrc_file = IRB.enum_for(:rc_file_generators).first.call(IRB::IRBRC_EXT)
 
-if irbrc_file && !File.exist?(irbrc_file)
+if irbrc_file && !File.exist?(irbrc_file) && File.exist?(File.dirname(irbrc_file))
   File.write irbrc_file, <<-EOT
 require 'irb/ext/save-history'
 require 'irb/completion'
